+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - metrics.clear() disables default metrics
 - cartridge role is permanent
 
+### Added
+- Luajit platform metrics
+
 ## [0.5.0] - 2020-09-18
 ### Added
 - Summary collector